[adress] Change buildong reforming condition
Change condition of changing of building string.
Move this condition to specified function to
parser_tools.py script.

Service that helps to get longitude and latitude
of buildings requires specific input format of
adress.
The conditionchecks if there are 'k' or 'K'
letters in building strigs. But it changes the
string only for cases where there is 'k', but not
for 'K'.

Signed-off-by: Nikita Sobolev <niksob97@mail.ru>