[parsing] Split address function to two different
There are some functions, that return more than
one value, that breaks logic of code. It is better
to split such functions.
Before there were 5 variables, that described address.
But it can be actually more (or less), than this
number of values. So, safe whole address to list
in database.
Upgrade longitude and latitude functions, because they
are depend on address function.

Signed-off-by: Nikita Sobolev <niksob97@mail.ru>