[alex-dev] Imroved parser tool and unsuccesfull attempt counter
In longitude and latitude was added case when the address wasn't
parsed and consequenlty long and lat can't be parsed. Also
adjusted counter of attempts to parse ad: there was difficult
chain of exceptions and I decided to remove one exception and to
use simple flag that is returned in download_data function and
which is zero when there is no errors and else it is  one.